requirements.txt
glfw==2.9.0
numpy==2.2.6
pyglm==2.8.2
PyOpenGL==3.1.9
